# Conflicts:
#	conf/routes

+++ app/models/services/EventService.scala
@@ -3,7 +3,7 @@ package models.services
 import forms.EventForm.EventData
 import models.{Event, User}
 import models.daos.{EventDAO, UserDAO}
-import play.api.libs.json.{JsObject, Json}
+import play.api.libs.json.Json
 
 import java.time.{Duration, LocalDateTime}
 import java.util.UUID
@@ -11,6 +11,13 @@ import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
 
 class EventService @Inject()(userDAO: UserDAO, eventDAO: EventDAO)(implicit ex: ExecutionContext) {
+
+  /**
+   * Извлекает список событий
+   *
+   * @return
+   */
+  def retrieveAll: Future[Seq[Event]] = eventDAO.getAll
 
   /**
    * Функция обрабатывает дату и время окончания события
@@ -63,12 +70,13 @@ class EventService @Inject()(userDAO: UserDAO, eventDAO: EventDAO)(implicit ex: 
    * @param newEndDateTime Обработанные дата и время окончания события
    * @return Результат выполнения операции и событие, которое было обновлено
    */
-  private def updateEventFunction(eventID: Long, eventData: EventData, members: List[UUID], newEndDateTime: LocalDateTime): Future[EventResult] = {
-    for {
-      memberUsers <- userDAO.findUsersByID(members).map(data => data.toList)
-      updatedEvent <-
-        eventDAO.update(Event(eventID, eventData.title, eventData.startDateTime, newEndDateTime, UUID.fromString(eventData.orgUserID), Some(Json.obj("users" -> Json.toJson(memberUsers))), eventData.itemID, eventData.description))
-    } yield EventUpdated(updatedEvent)
+  private def updateEventFunction(eventID: Long, eventData: EventData, members: List[UUID], newEndDateTime: LocalDateTime, currentUser: User): Future[EventResult] = {
+    if (currentUser.id == UUID.fromString(eventData.orgUserID) || currentUser.role.contains("Admin"))
+      for {
+        memberUsers <- userDAO.findUsersByID(members).map(data => data.toList)
+        updatedEvent <- eventDAO.update(Event(eventID, eventData.title, eventData.startDateTime, newEndDateTime, UUID.fromString(eventData.orgUserID), Some(Json.obj("users" -> Json.toJson(memberUsers))), eventData.itemID, eventData.description))
+      } yield EventUpdated(updatedEvent)
+    else Future.successful(EventCreatedByAnotherUser("update"))
   }
 
   /**
@@ -76,10 +84,11 @@ class EventService @Inject()(userDAO: UserDAO, eventDAO: EventDAO)(implicit ex: 
    *
    * @param eventID ID события
    * @param eventData Данные о событии
-   * @param members ID пользователей
+   * @param members ID пользователей-участников
+   * @param currentUser Объект данных авторизованного пользователей
    * @return Результат выполнения операции и событие, которое было обновлено
    */
-  def updateEvent(eventID: Long, eventData: EventData, members: List[UUID]): Future[EventResult] = {
+  def updateEvent(eventID: Long, eventData: EventData, members: List[UUID], currentUser: User): Future[EventResult] = {
     val newEndDateTime: LocalDateTime = newEventDateTimeBuilder(eventData.endDateTime)
 
     val compareDateTimeValue = eventData.startDateTime compareTo eventData.endDateTime
@@ -92,24 +101,34 @@ class EventService @Inject()(userDAO: UserDAO, eventDAO: EventDAO)(implicit ex: 
           if (eventWithID.startDateTime != eventData.startDateTime && eventWithID.endDateTime != eventData.endDateTime) {
             eventDAO.getByDateTime(eventData.startDateTime, eventData.endDateTime).flatMap {
               case Some(eventWithDateTime) =>
-                if(eventWithDateTime.id == eventID) updateEventFunction(eventID, eventData, members, newEndDateTime)
+                if (eventWithDateTime.id == eventID)
+                  updateEventFunction(eventID, eventData, members, newEndDateTime, currentUser)
                 else Future.successful(EventAlreadyExists)
-              case None => updateEventFunction(eventID, eventData, members, newEndDateTime)
+              case None => updateEventFunction(eventID, eventData, members, newEndDateTime, currentUser)
             }
-          } else updateEventFunction(eventID, eventData, members, newEndDateTime)
+          } else updateEventFunction(eventID, eventData, members, newEndDateTime, currentUser)
         case None =>
           Future.successful(EventNotFound)
       }
     }
   }
 
-  def deleteEvent(eventID: Long) : Future[EventResult] = {
+  /**
+   * Обрабатывает удаление данных события
+   *
+   * @param eventID ID события, которое необходимо удалить
+   * @param currentUser Объект данных авторизованного пользователя
+   * @return Результат выполнения операции
+   */
+  def deleteEvent(eventID: Long, currentUser: User) : Future[EventResult] = {
     eventDAO.getByID(eventID).flatMap {
-      case Some(_) =>
-        eventDAO.delete(eventID).flatMap {delResult =>
-          if (delResult) Future.successful(EventDeleted)
-          else Future.successful(Error("Произошла ошибка при удалении события!"))
-        }
+      case Some(eventData) =>
+        if (currentUser.id == eventData.orgUserId || currentUser.role.contains("Admin"))
+          eventDAO.delete(eventID).flatMap { delResult =>
+            if (delResult) Future.successful(EventDeleted)
+            else Future.successful(EventDeleteError("Произошла ошибка при удалении события!"))
+          }
+        else Future.successful(EventCreatedByAnotherUser("delete"))
       case None => Future.successful(EventNotFound)
     }
   }
@@ -127,5 +146,6 @@ case object EventNotFound extends EventResult
 case object EventDeleted extends EventResult
 
 case class EventCreated(event: Event) extends EventResult
+case class EventCreatedByAnotherUser(action: String) extends EventResult
 case class EventUpdated(event: Event) extends EventResult
-case class Error(msg: String) extends EventResult
+case class EventDeleteError(msg: String) extends EventResult
